Allow use of `local_` to adjust regions for loops. (#35)
This patch changes the local stack region behavior for the body of for loops,
and the body and condition of while loops.  Now, the user may request that no
new stack region is created for these positions, and instead local allocations
happen in the region containing the loop.

This is of limited use in isolation, but is useful with the coming "let mutable"
patch, where we want to allow locals in a loop body to "escape" to mutable
variables declared in the scope containing the loop.

added new lambda construct Lmutvar (#9954)

added new construct Lmutlet and removed Variable attribute

refactorized Llet and Lmutlet printing

Simplified some computations for mutable variable cases

Fixed an overlooked transformation to Lmutvar and added cautionnary comment

Fixups for backported PR (mostly in translcomprehension)

Add the "let mutable" extension.

This extension allows let-bound variables to be marked as mutable, as in:

let add_55 y =
  let mutable x = y in
  for i = 1 to 10 do
    x <- x + i
  done;
  x

Mutable variables are like "ref"s, but don't give you a way to refer to the ref
cell itself - all uses refer to the stored value.

The typechecker ensures that mutable variables do not escape the scope in which
they are defined.  This static check ensures, roughly, that the variable can
be implemented efficiently as a register, without allocation.  Additionally, it
allows for local values to be stored in mutable variables.

Rework `let mutable` to track mutability of lets, not patterns